refactor(api): move constant CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING to src
