+++ api/scripts/prod/compute-participation-results.js
@@ -14,8 +14,8 @@ const { SHARED } = CampaignParticipationStatuses;
 import _ from 'lodash';
 
 import { disconnect, knex } from '../../db/knex-database-connection.js';
-import { constants } from '../../lib/infrastructure/constants.js';
 import * as placementProfileService from '../../src/shared/domain/services/placement-profile-service.js';
+import { CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING } from '../../src/shared/infrastructure/constants.js';
 import * as competenceRepository from '../../src/shared/infrastructure/repositories/competence-repository.js';
 import { PromiseUtils } from '../../src/shared/infrastructure/utils/promise-utils.js';
 
@@ -98,7 +98,7 @@ async function _getCampaignParticipationChunks(campaign) {
     .join('campaigns', 'campaign-participations.campaignId', 'campaigns.id')
     .where({ campaignId: campaign.id, status: SHARED, pixScore: null, 'campaigns.type': 'ASSESSMENT' })
     .orWhere({ campaignId: campaign.id, status: SHARED, isCertifiable: null, 'campaigns.type': 'PROFILES_COLLECTION' });
-  return _.chunk(campaignParticipations, constants.CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING);
+  return _.chunk(campaignParticipations, CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING);
 }
 
 async function _computeResults(skillIds, competences, campaignParticipations) {
