+++ api/src/shared/infrastructure/constants.js
@@ -4,4 +4,6 @@ const CONCURRENCY_HEAVY_OPERATIONS = config.infra.concurrencyForHeavyOperations;
 
 const ORGANIZATION_LEARNER_CHUNK_SIZE = config.infra.chunkSizeForOrganizationLearnerDataProcessing;
 
-export { CONCURRENCY_HEAVY_OPERATIONS, ORGANIZATION_LEARNER_CHUNK_SIZE };
+const CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING = config.infra.chunkSizeForCampaignResultProcessing;
+
+export { CHUNK_SIZE_CAMPAIGN_RESULT_PROCESSING, CONCURRENCY_HEAVY_OPERATIONS, ORGANIZATION_LEARNER_CHUNK_SIZE };
